% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimest.cv.R
\name{dimest.cv}
\alias{dimest.cv}
\title{Estimate dimensionality by crossvalidation}
\usage{
dimest.cv(dmat,epsratio=.001,mds.control=list(type="ordinal",itmax=1000))
}
\arguments{
\item{dmat}{m x n matrix of n vectorized dissimilarity matrices
epsratio convergence criterion. Defaults to .001
mds.control list of smacofSym parameters: type, defaults to "ordinal"; itmax, defaults to 1000}
}
\value{
ndim estimated number of dimensions
cv.cor vector of mean crossvalidated correlations between mds configurations and left out data
}
\description{
Estimate dimensionality by change in crossvalidated multidimensional scaling fit
}
\details{
Defines identity matrix of dimensions n x n. Reproduces functionality of identically named MATLAB function.
}
\examples{
set.seed(1)
dat <- matrix(rnorm(200),100,2)
dmat <- as.vector(dist(dat))
dmat2 <- replicate(2,dmat+rnorm(4950))
dimcvfit <- dimest.cv(dmat2)
}
\seealso{
\code{\link{smacofSym}} \code{\link{dimest}} \code{\link{dimest.R2}}
}

