% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimest.R2.R
\name{dimest.R2}
\alias{dimest.R2}
\title{Estimate dimensionality by R-squared}
\usage{
dimest.R2(dmat,epsratio=.1,mds.control=list(type="ordinal",itmax=1000))
}
\arguments{
\item{dmat}{vectorized dissimilarity matrix
epsratio convergence ratio, defaults to .1
mds.control list of smacofSym parameters: type, defaults to "ordinal"; itmax, defaults to 1000}
}
\value{
ndim estimated number of dimensions
r2 R-squared of configuration with ndim dimensions
}
\description{
Estimate dimensionality by change in multdimensional scaling R-squared
}
\details{
Defines identity matrix of dimensions n x n. Reproduces functionality of identically named MATLAB function.
}
\examples{
set.seed(1)
dat <- matrix(rnorm(200),100,2)
dmat <- as.vector(dist(dat))
dimR2fit <- dimest.R2(dmat)
}
\seealso{
\code{\link{smacofSym}} \code{\link{dimest}} \code{\link{dimest.cv}}
}

