% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye.R
\name{eye}
\alias{eye}
\title{Identity matrix}
\usage{
eye(n)
}
\arguments{
\item{n}{size of matrix}
}
\value{
m an n x n identity matrix
}
\description{
Define identity matrix
}
\details{
Defines identity matrix of dimensions n x n. Reproduces functionality of identically named MATLAB function.
}
\examples{
eye(5)
}

