% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerTriangle.R
\name{lowerTriangle}
\alias{lowerTriangle}
\title{Lower triangle of matrix}
\usage{
lowerTriangle(m)
}
\arguments{
\item{m}{a square matrix}
}
\value{
v a vector containing the lower triangular
}
\description{
Extract and vectorize lower triangular portion of square matrix
}
\examples{
mat <- as.matrix(dist(1:10))
lowerTriangle(mat)
}

